Move `PaginatedTable` types out on its own file
