+++ packages/webapp/src/components/lists/ReportList/reports/RequestReport/useRequestReportColumnsOld.tsx
@@ -7,7 +7,7 @@ import { IDropdownOption } from '@fluentui/react'
 import { useMemo } from 'react'
 import { CustomDateRangeFilter } from '~components/ui/CustomDateRangeFilter'
 import { CustomTextFieldFilter } from '~components/ui/CustomTextFieldFilter'
-import { IPaginatedTableColumn } from '~components/ui/PaginatedTable'
+import { IPaginatedTableColumn } from '~components/ui/PaginatedTable/types'
 import { useLocale } from '~hooks/useLocale'
 import { Namespace, useTranslation } from '~hooks/useTranslation'
 import styles from '../../index.module.scss'
