+++ packages/webapp/src/components/ui/PaginatedTable/types.ts
@@ -1,0 +1,30 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+
+import type { StandardComponentProps } from '~types/StandardFC'
+
+export interface IPaginatedTableColumn {
+	key: string
+	name?: string
+	headerClassName?: string
+	itemClassName?: string
+	fieldName?: string | Array<string>
+	onRenderColumnHeader?: (key: string, name: string, index: number) => JSX.Element | string
+	onRenderColumnItem?: (item: any, index: number) => JSX.Element | JSX.Element[] | string
+}
+
+export interface PaginatedTableProps<T> extends StandardComponentProps {
+	list: T[]
+	itemsPerPage: number
+	columns: IPaginatedTableColumn[]
+	tableClassName?: string
+	headerRowClassName?: string
+	bodyRowClassName?: string
+	paginatorContainerClassName?: string
+	overFlowActiveClassName?: string
+	isMD?: boolean
+	isLoading?: boolean
+	onPageChange?: (items: T[], currentPage: number) => void
+}
